requirement.txt
loguru
fpdf2
